moved block viewer and all dependencies to a seperate folder
